merge(tests): unify generative tests into a single file :twisted_rightward_arrows:
