--- compiler/json_test.go
@@ -1,43 +1,0 @@
-/*
- * Copyright 2017 Workiva
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package compiler_test
-
-import (
-	"testing"
-	"time"
-
-	"github.com/Workiva/frugal/compiler"
-	"github.com/Workiva/frugal/compiler/globals"
-)
-
-func TestValidJsonFrugalCompiler(t *testing.T) {
-	defer globals.Reset()
-	nowBefore := globals.Now
-	defer func() {
-		globals.Now = nowBefore
-	}()
-	globals.Now = time.Date(2015, 11, 24, 0, 0, 0, 0, time.UTC)
-
-	options := compiler.Options{
-		File:    frugalGenFile,
-		Gen:     "json:indent",
-		Out:     outputDir,
-		Delim:   delim,
-		Recurse: true,
-	}
-	suite := ComparisonList{
-		{"frugal.json", "frugal.json"},
-	}
-	suite.Run(t, options)
-}
